Extract shared vector geometry stage from buildVectorGltfFromMVT
